move get-colorway util to resolvers/utils
